backend/requirements.txt
# backend/requirements.txt

fastapi>=0.100.0                 # The FastAPI framework
uvicorn[standard]>=0.20.0        # ASGI server with standard extras (for performance and --reload)
sqlalchemy>=2.0.0                # Database ORM
psycopg2-binary>=2.9.0           # PostgreSQL adapter (use psycopg2 if you need to build from source)
python-dotenv>=1.0.0             # For loading .env files
pydantic[email]>=2.0.0           # Data validation (email extra is optional but common)

# Add any other specific dependencies your application might need
